Add VIsa and IBox payments

+++ Gamestore/GameStore.Application/Dtos/OrderInformation.cs
@@ -1,10 +1,10 @@
 ﻿namespace GameStore.Application.Dtos;
 
-public class OrderInformation(Guid orderId, DateTime creationDate, double sum)
+public class OrderInformation(Guid orderId, DateTime creationDate, int sum)
 {
     public Guid OrderId { get; set; } = orderId;
 
     public DateTime CreationDate { get; set; } = creationDate;
 
-    public double Sum { get; set; } = sum;
+    public int Sum { get; set; } = sum;
 }
